+++ libs/platform/src/lib/table/components/table-view-settings-dialog/grouping/grouping.component.ts
@@ -23,6 +23,7 @@ export interface SettingsGroupDialogResultData {
 }
 
 const NOT_GROUPED_OPTION_VALUE = null;
+const INITIAL_DIRECTION = SortDirection.ASC;
 
 @Component({
     templateUrl: './grouping.component.html',
@@ -30,12 +31,6 @@ const NOT_GROUPED_OPTION_VALUE = null;
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class GroupingComponent implements Resettable {
-    /** Initially active direction */
-    initialDirection: SortDirection = SortDirection.ASC;
-
-    /** Initially active field */
-    initialField: string = null;
-
     /** Current selected direction */
     direction: SortDirection;
 
@@ -61,18 +56,16 @@ export class GroupingComponent implements Resettable {
     constructor(public dialogRef: DialogRef) {
         const data: SettingsGroupDialogData = this.dialogRef.data;
 
-        this.initialDirection = data.direction || this.initialDirection;
-        this.initialField = data.field || this.initialField;
         this.columns = data.columns || [];
 
-        this.direction = this.initialDirection;
-        this.field = this.initialField;
+        this.direction = data.direction ?? INITIAL_DIRECTION;
+        this.field = data.field ?? NOT_GROUPED_OPTION_VALUE;
     }
 
     /** Reset changes to the initial state */
     reset(): void {
-        this.direction = this.initialDirection;
-        this.field = this.initialField;
+        this.direction = INITIAL_DIRECTION;
+        this.field = NOT_GROUPED_OPTION_VALUE;
         this._isResetAvailableSubject$.next(false);
     }
 
@@ -102,7 +95,7 @@ export class GroupingComponent implements Resettable {
     /** @hidden */
     _onModelChange(): void {
         // Use this coercion cause fd-radio-button triggers extra ngModelChange events on initial phase
-        const isInitialDiffers = this.initialDirection !== this.direction || this.initialField !== this.field;
+        const isInitialDiffers = this.direction !== INITIAL_DIRECTION || this.field !== NOT_GROUPED_OPTION_VALUE;
         this._isResetAvailableSubject$.next(isInitialDiffers);
     }
 }